Split close-reason optimistic tests to keep file under 1000-line limit (Fixes #204)
