+++ src/state/issues_tests_close_delete.rs
@@ -5,7 +5,7 @@
 //! failure clears pending + sets scoped error, stale-mutation-id guards,
 //! scope-mismatch guards.
 
-use crate::domain::{Issue, IssueDetail, IssueState, IssueStateReason, RepositoryId};
+use crate::domain::{Issue, IssueDetail, IssueState, RepositoryId};
 use crate::state::AppState;
 use crate::state::{
     AppEvent, InlineState, IssueDeleteConfirmState, IssueFocus, IssueLifecycleMutationPending,
@@ -238,119 +238,6 @@ fn issue_closed_updates_list_and_detail_state() {
     assert!(
         notice.is_some_and(|n| n.contains("Closed issue #1")),
         "draft notice should mention closing, got {notice:?}"
-    );
-}
-
-#[test]
-fn issue_closed_with_reason_sets_optimistic_state_reason() {
-    let mut state = issues_state_with_list("repo-1");
-    state.issues_state.issue_detail = Some(make_detail(1, "I_1"));
-    let scope = RepositoryId("repo-1".to_string());
-    state.issues_state.close_mutation_pending = Some(IssueLifecycleMutationPending {
-        scope_repo_id: scope.clone(),
-        mutation_id: 42,
-        issue_number: 1,
-        node_id: Some("I_1".to_string()),
-        close_reason: Some(crate::domain::CloseReason::NotPlanned),
-        duplicate_of: None,
-    });
-    let state = state.apply(AppEvent::IssueClosed {
-        scope_repo_id: scope,
-        issue_number: 1,
-        mutation_id: 42,
-        close_reason: Some(crate::domain::CloseReason::NotPlanned),
-        duplicate_of: None,
-    });
-    let list_issue = state.issues_state.issues().iter().find(|i| i.number == 1);
-    assert!(
-        list_issue.is_some_and(|i| i.state_reason == Some(IssueStateReason::NotPlanned)),
-        "list row state_reason should be NotPlanned after close-with-reason"
-    );
-    let detail = state.issues_state.issue_detail.as_ref();
-    assert!(
-        detail.is_some_and(|d| d.state_reason == Some(IssueStateReason::NotPlanned)),
-        "detail state_reason should be NotPlanned after close-with-reason"
-    );
-}
-
-#[test]
-fn issue_closed_duplicate_sets_optimistic_duplicate_reason() {
-    let mut state = issues_state_with_list("repo-1");
-    state.issues_state.close_mutation_pending = Some(IssueLifecycleMutationPending {
-        scope_repo_id: RepositoryId("repo-1".to_string()),
-        mutation_id: 42,
-        issue_number: 1,
-        node_id: Some("I_1".to_string()),
-        close_reason: Some(crate::domain::CloseReason::Duplicate),
-        duplicate_of: Some(2),
-    });
-    let state = state.apply(AppEvent::IssueClosed {
-        scope_repo_id: RepositoryId("repo-1".to_string()),
-        issue_number: 1,
-        mutation_id: 42,
-        close_reason: Some(crate::domain::CloseReason::Duplicate),
-        duplicate_of: Some(2),
-    });
-    let list_issue = state.issues_state.issues().iter().find(|i| i.number == 1);
-    assert!(
-        list_issue.is_some_and(|i| i.state_reason == Some(IssueStateReason::Duplicate)),
-        "list row state_reason should be Duplicate"
-    );
-
-    // The reducer updates both list row and detail when issue_detail is loaded.
-    let mut state = state;
-    state.issues_state.issue_detail = Some(make_detail(1, "I_1"));
-    let scope = RepositoryId("repo-1".to_string());
-    state.issues_state.close_mutation_pending = Some(IssueLifecycleMutationPending {
-        scope_repo_id: scope.clone(),
-        mutation_id: 43,
-        issue_number: 1,
-        node_id: Some("I_1".to_string()),
-        close_reason: Some(crate::domain::CloseReason::Duplicate),
-        duplicate_of: Some(2),
-    });
-    let state = state.apply(AppEvent::IssueClosed {
-        scope_repo_id: scope,
-        issue_number: 1,
-        mutation_id: 43,
-        close_reason: Some(crate::domain::CloseReason::Duplicate),
-        duplicate_of: Some(2),
-    });
-    let detail = state.issues_state.issue_detail.as_ref();
-    assert!(
-        detail.is_some_and(|d| d.state_reason == Some(IssueStateReason::Duplicate)),
-        "detail state_reason should be Duplicate"
-    );
-}
-
-#[test]
-fn issue_closed_plain_close_defaults_to_completed_reason() {
-    let mut state = issues_state_with_list("repo-1");
-    state.issues_state.issue_detail = Some(make_detail(1, "I_1"));
-    state.issues_state.close_mutation_pending = Some(IssueLifecycleMutationPending {
-        scope_repo_id: RepositoryId("repo-1".to_string()),
-        mutation_id: 42,
-        issue_number: 1,
-        node_id: Some("I_1".to_string()),
-        close_reason: None,
-        duplicate_of: None,
-    });
-    let state = state.apply(AppEvent::IssueClosed {
-        scope_repo_id: RepositoryId("repo-1".to_string()),
-        issue_number: 1,
-        mutation_id: 42,
-        close_reason: None,
-        duplicate_of: None,
-    });
-    let list_issue = state.issues_state.issues().iter().find(|i| i.number == 1);
-    assert!(
-        list_issue.is_some_and(|i| i.state_reason == Some(IssueStateReason::Completed)),
-        "plain close should default state_reason to Completed"
-    );
-    let detail = state.issues_state.issue_detail.as_ref();
-    assert!(
-        detail.is_some_and(|d| d.state_reason == Some(IssueStateReason::Completed)),
-        "detail state_reason should default to Completed on plain close"
     );
 }
